refactor(ch10): merge card.py into executor.py
build_agent_card() moves into executor.py alongside
LLMAgentA2AExecutor, and a2a/server/card.py is deleted.

Each executor variant owning its own card-building function,
colocated, is about to become the actual organizing principle once
#814's streaming executor lands (streaming_executor.py will house
both a StreamingLLMAgentA2AExecutor and its own
build_streaming_agent_card()) -- a shared card.py would just become a
dumping ground for loosely-related build_X_agent_card() functions for
different executors. This isn't the same tension Decision 14 rejected
(a composite class blending static-identity and per-request-
translation lifetimes into one abstraction): build_agent_card() stays
a free function, not a method on LLMAgentA2AExecutor, so it's still
never part of the SDK's AgentExecutor contract -- just colocated in
the same file as the executor whose behavior it describes.

tests/a2a/server/test_card.py stays a separate test file despite the
source merge, matching the existing agent/llm_agent.py precedent
(one large source file, tests split by concept across test_agent.py /
test_task_handler.py).
