+++ src/llm_agents_from_scratch/a2a/server/__init__.py
@@ -1,7 +1,6 @@
 """A2A server-side module — serving an LLMAgent as an A2A peer."""
 
-from .card import build_agent_card
-from .executor import LLMAgentA2AExecutor
+from .executor import LLMAgentA2AExecutor, build_agent_card
 
 __all__ = [
     "LLMAgentA2AExecutor",
